nepse_ai_trading/requirements.txt
# NEPSE AI Trading Bot - Dependencies
# =====================================

# Core Data Processing
pandas==2.2.0
numpy==1.26.3
pandas-ta==0.3.14b0

# NEPSE Unofficial API
# Option 1: Install from local folder (if you downloaded it)
# pip install -e ../NepseUnofficialApi
# Option 2: Install from GitHub (default)
nepse @ git+https://github.com/basic-bgnr/NepseUnofficialApi

# HTTP (required by NepseUnofficialApi)
httpx==0.26.0
tqdm==4.66.1

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP & API
requests==2.31.0
aiohttp==3.9.1

# Web Scraping
playwright==1.41.0

# AI Integration
openai==1.10.0
tiktoken==0.5.2

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
jinja2==3.1.3
python-multipart==0.0.6

# Authentication (for future SaaS)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Notifications
python-telegram-bot==20.7
aiosmtplib==3.0.1

# Scheduler
apscheduler==3.10.4
pytz==2024.1

# Environment & Config
python-dotenv==1.0.0

# Logging & Monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Development
black==24.1.0
isort==5.13.2
mypy==1.8.0

# Backtesting & Visualization
matplotlib==3.8.2
